+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## [none]
+## [3.2.6]
 
+- Enhanced the `ABridge` contract architecture.
+  - Adding modular handler contracts for assets and messages management.
+  - Added `Batcher` contract for batched execution during redemption.
+  - Added arbitrary messages handler.
+  - Removed ERC721 and ERC1155 handlers.
+- Fixed encoding of locktime in `TxParser` library.
 - Clarified CI sequence.
 
 ## [3.2.5]